matches: new module, move single_match lint there
